+++ src/main/java/com/example/gaegizo/domain/memo/domain/Memo.java
@@ -1,0 +1,19 @@
+package com.example.gaegizo.domain.memo.domain;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+import javax.persistence.Entity;
+
+//@Entity
+//@Getter
+//@NoArgsConstructor
+//@AllArgsConstructor
+//@Builder
+//public class Memo {
+//
+//    private Long id;
+//    private String content;
+//}
