+++ src/main/resources/data.sql
@@ -1,6 +1,6 @@
 
-INSERT INTO INTEREST_JOB(EDUCATION, AVERAGE_SALARY, CAREER, JOB, JOB_NUMBER, PREFERENTIAL_TREATMENT, REGION, SALARY, SECTORS, STATISTICS, TITLE, WELFARE, WORK_TYPE, DEADLINE) VALUES('학력 무관', 2800, '경력 무관', '운송·운수·물류','42377191', '해당 분야 전공자 및 유경험자 우대', '경기도 성남시', '면접 후 결정', 'IT웹 서비스', 80, '이자경수학 아이파크관에서 고등 수학선생님 구인합니다', '4대 보험', '정규직', '상시채용')
+INSERT INTO INTEREST_JOB(EDUCATION, AVERAGE_SALARY, CAREER, JOB, JOB_NUMBER, PREFERENTIAL_TREATMENT, REGION, SALARY, SECTORS, STATISTICS, TITLE, WELFARE, WORK_TYPE, DEADLINE) VALUES('학력 무관', 2800, '경력 무관', '운송·운수·물류','42320380', '해당 분야 전공자 및 유경험자 우대', '경기도 성남시', '면접 후 결정', 'IT웹 서비스', 80, '이자경수학 아이파크관에서 고등 수학선생님 구인합니다', '4대 보험', '정규직', '상시채용')
 INSERT INTO INTEREST_JOB(EDUCATION, AVERAGE_SALARY, CAREER, JOB, JOB_NUMBER, PREFERENTIAL_TREATMENT, REGION, SALARY, SECTORS, STATISTICS, TITLE, WELFARE, WORK_TYPE, DEADLINE) VALUES('학사 이상', 3200, '경력 1년이상', '운송·운수·물류','42377191', '해당 분야 전공자 및 유경험자 우대', '경기도 안산시', '면접 후 결정', 'IT웹 서비스', 10, '이자경수학 아이파크관에서 고등 수학선생님 구인합니다', '4대 보험', '정규직','3월말')
-INSERT INTO INTEREST_JOB(EDUCATION, AVERAGE_SALARY, CAREER, JOB, JOB_NUMBER, PREFERENTIAL_TREATMENT, REGION, SALARY, SECTORS, STATISTICS, TITLE, WELFARE, WORK_TYPE, DEADLINE) VALUES('석사 이상', 3000, '경력 3년이상', '운송·운수·물류','42377191', '해당 분야 전공자 및 유경험자 우대', '경기도 용인시', '면접 후 결정', 'IT웹 서비스', 180, '이자경수학 아이파크관에서 고등 수학선생님 구인합니다', '4대 보험', '정규직','3/18')
-INSERT INTO INTEREST_JOB(EDUCATION, AVERAGE_SALARY, CAREER, JOB, JOB_NUMBER, PREFERENTIAL_TREATMENT, REGION, SALARY, SECTORS, STATISTICS, TITLE, WELFARE, WORK_TYPE, DEADLINE) VALUES('학사 이상', 2600, '경력 무관', '운송·운수·물류','42377191', '해당 분야 전공자 및 유경험자 우대', '인천광역시 부평구', '면접 후 결정', 'IT웹 서비스', 60, '이자경수학 아이파크관에서 고등 수학선생님 구인합니다', '4대 보험', '정규직', '3/12')
-INSERT INTO INTEREST_JOB(EDUCATION, AVERAGE_SALARY, CAREER, JOB, JOB_NUMBER, PREFERENTIAL_TREATMENT, REGION, SALARY, SECTORS, STATISTICS, TITLE, WELFARE, WORK_TYPE, DEADLINE) VALUES('학력 무관', 2800, '경력 무관', '운송·운수·물류','42377191', '해당 분야 전공자 및 유경험자 우대', '서울특별시 강남구', '면접 후 결정', 'IT웹 서비스', 20, '이자경수학 아이파크관에서 고등 수학선생님 구인합니다', '4대 보험', '정규직','상시채용')
+INSERT INTO INTEREST_JOB(EDUCATION, AVERAGE_SALARY, CAREER, JOB, JOB_NUMBER, PREFERENTIAL_TREATMENT, REGION, SALARY, SECTORS, STATISTICS, TITLE, WELFARE, WORK_TYPE, DEADLINE) VALUES('석사 이상', 3000, '경력 3년이상', '운송·운수·물류','42349073', '해당 분야 전공자 및 유경험자 우대', '경기도 용인시', '면접 후 결정', 'IT웹 서비스', 180, '이자경수학 아이파크관에서 고등 수학선생님 구인합니다', '4대 보험', '정규직','3/18')
+INSERT INTO INTEREST_JOB(EDUCATION, AVERAGE_SALARY, CAREER, JOB, JOB_NUMBER, PREFERENTIAL_TREATMENT, REGION, SALARY, SECTORS, STATISTICS, TITLE, WELFARE, WORK_TYPE, DEADLINE) VALUES('학사 이상', 2600, '경력 무관', '운송·운수·물류','42358928', '해당 분야 전공자 및 유경험자 우대', '인천광역시 부평구', '면접 후 결정', 'IT웹 서비스', 60, '이자경수학 아이파크관에서 고등 수학선생님 구인합니다', '4대 보험', '정규직', '3/12')
+INSERT INTO INTEREST_JOB(EDUCATION, AVERAGE_SALARY, CAREER, JOB, JOB_NUMBER, PREFERENTIAL_TREATMENT, REGION, SALARY, SECTORS, STATISTICS, TITLE, WELFARE, WORK_TYPE, DEADLINE) VALUES('학력 무관', 2800, '경력 무관', '운송·운수·물류','42370652', '해당 분야 전공자 및 유경험자 우대', '서울특별시 강남구', '면접 후 결정', 'IT웹 서비스', 20, '이자경수학 아이파크관에서 고등 수학선생님 구인합니다', '4대 보험', '정규직','상시채용')
